server/serverInventory/requirements.txt
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
pip-chill==1.0.3
pydantic==2.6.4
pymysql==1.1.0
flask==3.0.2
